LayoutTests/fast/regex/alternative-length-miscalculation.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description(
"This page tests for length miscalculations in regular expression processing."
);

var re = /b|[^b]/g;
re.lastIndex = 1;
shouldBe("re.exec('a')", "null");

var re2 = /[^a]|ab/;
shouldBe("re2.test('')", "false");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
